requirements.txt
agate==1.9.1
aiosqlite==0.22.1
alembic==1.18.4
amplitude-analytics==1.2.3
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
apprise==1.9.9
asgi-lifespan==2.1.0
asyncpg==0.31.0
attrs==26.1.0
babel==2.18.0
beartype==0.22.9
cachetools==7.0.5
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
cloudpickle==3.1.2
colorama==0.4.6
coolname==4.1.0
cronsim==2.7
cryptography==46.0.6
cyclopts==4.10.1
daff==1.4.2
dateparser==1.4.0
dbt-adapters==1.22.10
dbt-common==1.37.3
dbt-core==1.11.7
dbt-extractor==0.6.0
dbt-postgres==1.10.0
dbt-protos==1.0.443
dbt-semantic-interfaces==0.9.0
deepdiff==8.6.2
docker==7.1.0
docstring_parser==0.17.0
docutils==0.22.4
exceptiongroup==1.3.1
fakeredis==2.34.1
fastapi==0.135.3
fsspec==2026.3.0
graphviz==0.21
greenlet==3.3.2
griffe==2.0.2
griffecli==2.0.2
griffelib==2.0.2
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
humanize==4.15.0
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.1
isodate==0.7.2
Jinja2==3.1.6
jinja2-humanize-extension==0.4.0
jsonpatch==1.33
jsonpointer==3.1.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
leather==0.4.1
lupa==2.6
Mako==1.3.10
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mashumaro==3.14
mdurl==0.1.2
more-itertools==10.8.0
msgpack==1.1.2
networkx==3.6.1
numpy==2.4.4
oauthlib==3.3.1
opentelemetry-api==1.40.0
orderly-set==5.5.0
orjson==3.11.8
packaging==26.0
pandas==3.0.1
parsedatetime==2.6
pathspec==0.12.1
pluggy==1.6.0
prefect==3.6.24
prometheus_client==0.24.1
protobuf==6.33.6
psycopg2-binary==2.9.11
py-key-value-aio==0.4.4
pycparser==3.0
pydantic==2.12.5
pydantic-extra-types==2.11.1
pydantic-settings==2.13.1
pydantic_core==2.41.5
pydocket==0.18.2
Pygments==2.20.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-json-logger==4.1.0
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2026.1.post1
PyYAML==6.0.3
readchar==4.2.1
redis==7.4.0
referencing==0.37.0
regex==2026.3.32
requests==2.33.1
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rich==14.3.3
rich-rst==1.3.2
rpds-py==0.30.0
ruamel.yaml==0.19.1
ruamel.yaml.clib==0.2.15
semver==3.0.4
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
snowplow-tracker==1.1.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.48
sqlparse==0.5.4
starlette==1.0.0
text-unidecode==1.3
toml==0.10.2
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
uncalled-for==0.2.0
urllib3==2.6.3
uvicorn==0.42.0
websockets==16.0
whenever==0.9.5
zipp==3.23.0
